Increase mcast_rate to 24Mbit/s
Broadcast frames (which are a special form of multicast frames) are used by
batman-adv to detect the loss rate between nodes. This is then converted to
TQ between originators.

The current mcast_rate of 12Mbit/s (QPSK, coding rate 1/2) is comparable to
MCS 1. Just slightly above the minimal possible MCS rate. Using a higher
rate might better reflect the unicast experience (which is often trying to
use a lot higher MCS rate).

24 Mbit/s (16-QAM, coding rate 1/2) is comparable with the medium 1x1 MCS
rate 3. This is hopefully a good compromise between reliability of
multicast transmissions and quality of the TQ measurement.