Disable ffho-autoupdater-wifi-fallback
It seems like this package can cause problems related to missing local-node
ip addresses:

  root@PL-Bahnhof11-Shamland:~# ip addr show dev local-node
  136: local-node@local-port: <BROADCAST,MULTICAST,M-DOWN> mtu 1500 qdisc noqueue state DOWN qlen 1000
      link/ether 16:41:95:40:f7:dc brd ff:ff:ff:ff:ff:ff

  root@PL-Bahnhof11-Shamland:~# ubus call network.interface.local_node status
  {
          "up": false,
          "pending": false,
          "available": true,
          "autostart": true,
          "dynamic": false,
          "proto": "static",
          "device": "local-node",
          "data": {

          },
          "errors": [
                  {
                          "subsystem": "interface",
                          "code": "DEVICE_CLAIM_FAILED"
                  }
          ]
  }

This breaks for example DNS because the clients can no longer reach the
anycast IP address of the AP. The package can also be disabled via:

  uci set autoupdater-wifi-fallback.settings.enabled='0'
  uci commit autoupdater-wifi-fallback

The underlying problem is unclear but it seems like often reconfiguring the
device using `/etc/init.d/network restart` can trigger this "randomly". And
this package is doing this when the device is considered offline.

Signed-off-by: Sven Eckelmann <sven@narfation.org>